+++ src/filetypes/bkpdf.h
@@ -114,8 +114,8 @@ class BKPDF : public BKDocument {
 	virtual int setRotation(int, bool bForce=false);
 
 	virtual bool isBookmarkable();
-	virtual void getBookmarkPosition(map<string, int>&);
-	virtual int setBookmarkPosition(map<string, int>&);
+	virtual void getBookmarkPosition(map<string, float>&);
+	virtual int setBookmarkPosition(map<string, float>&);
 	virtual int getFastImageStatus();
 	virtual float getCurrentZoom();
 	virtual int getOutlineType();
